requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==4.1.0
boto3==1.28.78
botocore==1.31.78
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.5
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
django-storages==1.14.2
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.2
packaging==23.2
payme-pkg==2.6.3
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.25.4
vine==5.0.0
wcwidth==0.2.8
zope.interface==6.0
